Renaming generic template.
