+++ src/main/java/marc/henrard/risq/model/calibration/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Copyright (C) 2016 - present by Marc Henrard.
+ */
+
+/**
+ * Package with classes related to the calibration of model parameters.
+ */
+package marc.henrard.risq.model.calibration;
+
