+++ src/main/java/marc/henrard/risq/model/rationalmulticurve/RationalOneFactorSimpleHWShapeTemplate.java
@@ -33,6 +33,7 @@ import com.opengamma.strata.math.impl.minimization.SingleRangeLimitTransform;
 import com.opengamma.strata.math.impl.minimization.UncoupledParameterTransforms;
 import com.opengamma.strata.pricer.DiscountFactors;
 
+import marc.henrard.risq.model.generic.SingleCurrencyModelTemplate;
 import marc.henrard.risq.model.generic.TimeMeasurement;
 
 /**
@@ -44,7 +45,7 @@ import marc.henrard.risq.model.generic.TimeMeasurement;
  */
 @BeanDefinition(factoryName = "of")
 public final class RationalOneFactorSimpleHWShapeTemplate 
-    implements RationalTemplate, ImmutableBean, Serializable  {
+    implements SingleCurrencyModelTemplate, ImmutableBean, Serializable  {
 
   private static final double LIMIT_0 = 1.0E-8;
   private static final ParameterLimitsTransform[] DEFAULT_TRANSFORMS;
