+++ src/main/java/marc/henrard/risq/model/rationalmulticurve/RationalTwoFactor2HWShapePlusCstTemplate.java
@@ -36,6 +36,7 @@ import com.opengamma.strata.math.impl.minimization.SingleRangeLimitTransform;
 import com.opengamma.strata.math.impl.minimization.UncoupledParameterTransforms;
 import com.opengamma.strata.pricer.DiscountFactors;
 
+import marc.henrard.risq.model.generic.SingleCurrencyModelTemplate;
 import marc.henrard.risq.model.generic.TimeMeasurement;
 
 /**
@@ -45,7 +46,7 @@ import marc.henrard.risq.model.generic.TimeMeasurement;
  */
 @BeanDefinition(factoryName = "of")
 public final class RationalTwoFactor2HWShapePlusCstTemplate
-    implements RationalTemplate, ImmutableBean, Serializable  {
+    implements SingleCurrencyModelTemplate, ImmutableBean, Serializable  {
 
   private static final int NB_PARAMETERS = 10;
   private static final double LIMIT_0 = 1.0E-8;
